Feature/studyroom admin

See merge request s09-webmobile1-sub2/S09P12A206!237

+++ backend/sswm/src/main/java/com/ground/sswm/dailyLog/model/dto/CalenderDto.java
@@ -8,8 +8,8 @@ import lombok.ToString;
 @Setter
 @ToString
 public class CalenderDto {
-    private int studyExp;
-    private int restTimeExp;
-    private int stretchExp;
+    private float studyExp;
+    private float restTimeExp;
+    private float stretchExp;
     private int studyTime;
 }
